requirements.txt
# generating strings
rstr
# generating images
pillow
